requirements.txt
Django==1.4.2
MySQL-python==1.2.3
peewee==2.0.2
gunicorn==0.16.1
django-auth-ldap==1.1.2
python-ldap==2.4.10
